Fix the signin issue
Fix the issue when users are still in signed in state, the recipe and fridge don't have the data fetched.
