+++ Frontend/components/store.js
@@ -1,7 +1,6 @@
 import { createStore } from 'state-pool';
 
 const initialState = {
-  status: 0,
   foodList: [],
   recipeList: [],
   signedIn: false,
